NEPT-509: Port from the support-2.2 - fix(Configuration & Maintenance) (#1225)
* fix(Configuration & Maintenance): NEPT-509 - Port from the support-2.2 to the master branch of the Module reset steps

*  fix(Configuration & Maintenance): NEPT-509 - Improve phpDoc and method names

* NEPT-509: Add the "throws" phpdoc in restoreInitialState
