+++ profiles/common/modules/features/nexteuropa_dgt_connector/tmgmt_poetry/tmgmt_poetry_mock/tmgmt_poetry_mock.install
@@ -52,8 +52,45 @@ function tmgmt_poetry_mock_install() {
  * Implements hook_uninstall().
  */
 function tmgmt_poetry_mock_uninstall() {
-  $translator = tmgmt_translator_load('tmgmt_poetry_test_translator');
-  if (!empty($translator)) {
-    entity_delete('tmgmt_translator', $translator->tid);
+  // Delete the tmgmt_poetry_test_translator.
+  // Note tmgmt_translator_load is not stable in uninstall cases; it returns
+  // FALSE while the translator exists in the DB.
+  // That is why we pass by EntityFieldQuery.
+  $query = new EntityFieldQuery();
+  $result = $query->entityCondition('entity_type', 'tmgmt_translator')
+    ->propertyCondition('name', 'tmgmt_poetry_test_translator')
+    ->execute();
+  if (!empty($result['tmgmt_translator'])) {
+    $controller = entity_get_controller('tmgmt_translator');
+    // Since the entity controller handles the deletion of the attached
+    // entities (messages, job items) we just need to invoke it directly.
+    $controller->delete(array_keys($result['tmgmt_translator']));
+  }
+
+  // Delete translation jobs related to the tmgmt_poetry_test_translator.
+  // Code is inspired by tmgmt_cron().
+  $query = new EntityFieldQuery();
+  $result = $query->entityCondition('entity_type', 'tmgmt_job')
+    ->propertyCondition('translator', 'tmgmt_poetry_test_translator')
+    ->execute();
+  if (!empty($result['tmgmt_job'])) {
+    $controller = entity_get_controller('tmgmt_job');
+    // Since the entity controller handles the deletion of the attached
+    // entities (messages, job items) we just need to invoke it directly.
+    $controller->delete(array_keys($result['tmgmt_job']));
+  }
+
+  // Cleans Mock files.
+  $field_query = new EntityFieldQuery();
+  $file_fids = $field_query->entityCondition('entity_type', 'file')
+    ->propertyCondition('filemime', 'application/xml', '=')
+    ->propertyCondition('uri', db_like(TMGMT_POETRY_MOCK_REQUESTS_PATH) . '%', 'LIKE')
+    ->execute();
+  if (isset($file_fids['file'])) {
+    $fids = array_keys($file_fids['file']);
+    $files = file_load_multiple($fids);
+    foreach ($files as $file) {
+      file_delete($file);
+    }
   }
 }
